GLOBAL-1544 reply document

+++ features/messenger/messenger-impl/src/main/java/io/fasthome/fenestram_messenger/messenger_impl/data/service/model/ContentResponse.kt
@@ -15,6 +15,6 @@ class ContentResponse(
     /**
      * Размер файла указанный в мегабайтах
      */
-    var size: Int,
+    var size: Float,
     var url: String,
 ) : Parcelable
